Changed *map and *addGraf to be one map only
Also: fixed typos and updated so now it also logs map on which commands were executed. Additionally changed code for *map to create additional <img> instead of changing url of backgroundImage and added fallback when you invoke *light command without number (then it resets like *map so it's consistent)

"first unnamed map" is a phenomenon which occurs when there is no messages on chat on loading the game. Game then doesn't send name of map on which you spawned, so easy implementation falls short here, because commands directly on nameless map would be seen on all maps, which is bad in case of *addGraf and *map

New code uses sessionStorage to store timestamps of messages connected to maps on which they were executed. sessionStorage is supported up to IE8 and it's so huge there is no need to check if it's available, as there is almost no chance users would run this code on such old browser (worst case scenario first "unnamed" map can be buggy). Cookies are out of option in this case due to relatively big size of data so sessionStorage is way to go.

There is one rare case I'm aware of in which this code might not work (I haven't strictly tested it):
user sees command on map A
user goes to map B
user quits web browser, but then immediately logs back, so that sessionStorage is deleted but chat doesn't recognize him as loged off to clear chat.
Then user will see objects from map A on map B

Fixing this issue would presumably require creating cookie on closing the page with extremely low expiration time (10s-3min, exact timing must be tested) and then on new load checking if it's set. It requires deep interference and testing and I don't think it's necessary to do it (at least for now).